+++ src/resources/views/widgets/avatar.blade.php
@@ -1,2 +1,1 @@
-<img src="{{ avatar_image_url($data, $size*2) }}" class="img-avatar img-avatar-{{ $size }}" style="width: {{ $size }}px;" @if($tooltip)title="{{ $tooltip }}"@endif>
-
+@if($url)<a href="{{ $url }}">@endif<img src="{{ avatar_image_url($data, $size*2) }}" class="img-avatar img-avatar-{{ $size }}" style="width: {{ $size }}px;" @if($tooltip)title="{{ $tooltip }}"@endif>@if($url)</a>@endif